move AggCall to agg mod
