+++ src/batch/src/executor/sort_agg.rs
@@ -19,10 +19,9 @@ use risingwave_common::catalog::{Field, Schema};
 use risingwave_common::error::{Result, RwError};
 use risingwave_common::util::iter_util::ZipEqFast;
 use risingwave_expr::agg::{
-    build as build_agg, create_sorted_grouper, BoxedAggState, BoxedSortedGrouper, EqGroups,
+    build as build_agg, create_sorted_grouper, AggCall, BoxedAggState, BoxedSortedGrouper, EqGroups,
 };
 use risingwave_expr::expr::{build_from_prost, BoxedExpression};
-use risingwave_expr::function::aggregate::AggCall;
 use risingwave_pb::batch_plan::plan_node::NodeBody;
 
 use crate::executor::{
