+++ src/stream/src/executor/over_window/mod.rs
@@ -501,7 +501,7 @@ mod tests {
     use risingwave_common::test_prelude::StreamChunkTestExt;
     use risingwave_common::types::DataType;
     use risingwave_common::util::sort_util::OrderType;
-    use risingwave_expr::function::aggregate::AggArgs;
+    use risingwave_expr::agg::AggArgs;
     use risingwave_expr::function::window::{Frame, WindowFuncCall, WindowFuncKind};
     use risingwave_storage::memory::MemoryStateStore;
     use risingwave_storage::StateStore;
