+++ src/stream/src/from_proto/global_simple_agg.rs
@@ -14,7 +14,7 @@
 
 //! Streaming Aggregators
 
-use risingwave_expr::function::aggregate::AggCall;
+use risingwave_expr::agg::AggCall;
 use risingwave_pb::stream_plan::SimpleAggNode;
 
 use super::agg_common::{
