+++ src/stream/src/from_proto/local_simple_agg.rs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use risingwave_expr::function::aggregate::AggCall;
+use risingwave_expr::agg::AggCall;
 use risingwave_pb::stream_plan::SimpleAggNode;
 
 use super::*;
